+++ src/test/01.nextcloudClient.test.ts
@@ -30,7 +30,7 @@ describe("NEXCLOUD-NODE-CLIENT", function() {
         }
     });
 
-    it("02 get quota", async () => {
+    it.only("02 get quota", async () => {
 
         let q;
         try {