+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+### Changed
+- refactoring: getQuota without webdav client
+
 ## 0.3.0 - 2019-03-21
 ### Changed
 - incompatible api change: proxy in client constructor has a new type